Update reporting of status
Previous version reported heating as being on when it is off. This is because data.dcb.device_on was being used to identify the heating status when it should be data.dcb.heating_on. Changes made to device() and online() to fix this.

You might consider this to be a feature rather than a bug,. In determineIfHolding(), the line messages.push(`The heating is${qualifier} on.`); in particular will now give the correct message from my perspective