Correct reporting of heating on status
The previous version would report that the heating was on when it wasn't (in smartheat-core/core/ThermostatService.js - determineIfHolding() - messages.push(`The heating is${qualifier} on.`); ). This is because data.dcb.device_on was being used in device() to determine if the heating is on, when data.dcb.heating_on should be used.